+++ tscore.cpp
@@ -9,12 +9,12 @@ TScore::TScore()
     TScore::setGoals(0, 0);
 }
 
-TScore::TScore(int goalsHome, int goalsGuest)
+TScore::TScore(short goalsHome, short goalsGuest)
 {
     TScore::setGoals(goalsHome, goalsGuest);
 }
 
-void TScore::setGoals(int goalsHome, int goalsGuest)
+void TScore::setGoals(short goalsHome, short goalsGuest)
 {
     if ( goalsHome >= 0 and goalsHome <= 100 and goalsGuest >= 0 and goalsGuest <= 100 )
     {
@@ -28,7 +28,7 @@ void TScore::setGoals(int goalsHome, int goalsGuest)
     }
 }
 
-static int calcPoints(int first, int second)
+static short calcPoints(short first, short second)
 {
     if (first == second)
         return 1;
@@ -38,12 +38,12 @@ static int calcPoints(int first, int second)
         return 0;
 }
 
-int TScore::getPointsHome()
+short TScore::getPointsHome()
 {
    return calcPoints(TScore::goalsHome, TScore::goalsGuest);
 }
 
-int TScore::getPointsGuest()
+short TScore::getPointsGuest()
 {
     return calcPoints(TScore::goalsGuest, TScore::goalsHome);
 }
